# ffmpeg-full: Add support for rav1e (faster AV1 encoding)