@tomberek 2020-09-06
Result of `nixpkgs-review pr 95188` [1](https://github.com/Mic92/nixpkgs-review)
<details>
  <summary>1 package marked as broken and skipped:</summary>
<br>- meguca
</details>
<details>
  <summary>9 packages failed to build:</summary>
<br>- liquidsoap
<br>- python27Packages.pytmx
<br>- python37Packages.baselines
<br>- python37Packages.gym
<br>- python37Packages.pytmx
<br>- python37Packages.roboschool
<br>- python38Packages.gym
<br>- python38Packages.pytmx
<br>- python38Packages.roboschool
</details>
<details>
  <summary>17 packages built:</summary>
<br>- ffmpeg-full
<br>- handbrake
<br>- hydron
<br>- kdenlive (kdeApplications.kdenlive)
<br>- manim
<br>- olive-editor
<br>- printrun
<br>- python27Packages.pydub
<br>- python27Packages.pyglet
<br>- python37Packages.moderngl-window
<br>- python37Packages.pydub
<br>- python37Packages.pyglet
<br>- python38Packages.moderngl-window
<br>- python38Packages.pydub
<br>- python38Packages.pyglet
<br>- soundkonverter
<br>- webcamoid
</details>